pyproject.toml
[project]
name = "cyber-persona"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
    "apscheduler>=3.11.0",
    "google-auth-oauthlib>=1.2.2",
    "google-auth>=2.40.0",
    "google-api-python-client>=2.169.0",
]

[project.scripts]
cp = "cyber_persona.__main__:main"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
]

[tool.uv]
package = true
